fxd bit masks to provide 64 players support

+++ src/game/server/eventhandler.cpp
@@ -17,7 +17,7 @@ void CEventHandler::SetGameServer(CGameContext *pGameServer)
 	m_pGameServer = pGameServer;
 }
 
-void *CEventHandler::Create(int Type, int Size, int Mask)
+void *CEventHandler::Create(int Type, int Size, int64_t Mask)
 {
 	if(m_NumEvents == MAX_EVENTS)
 		return 0;
