+++ src/game/server/eventhandler.h
@@ -3,6 +3,8 @@
 #ifndef GAME_SERVER_EVENTHANDLER_H
 #define GAME_SERVER_EVENTHANDLER_H
 
+#include <stdint.h>
+
 //
 class CEventHandler
 {
@@ -12,7 +14,7 @@ class CEventHandler
 	int m_aTypes[MAX_EVENTS]; // TODO: remove some of these arrays
 	int m_aOffsets[MAX_EVENTS];
 	int m_aSizes[MAX_EVENTS];
-	int m_aClientMasks[MAX_EVENTS];
+	int64_t m_aClientMasks[MAX_EVENTS];
 	char m_aData[MAX_DATASIZE];
 
 	class CGameContext *m_pGameServer;
@@ -24,7 +26,7 @@ public:
 	void SetGameServer(CGameContext *pGameServer);
 
 	CEventHandler();
-	void *Create(int Type, int Size, int Mask = -1);
+	void *Create(int Type, int Size, int64_t Mask = -1);
 	void Clear();
 	void Snap(int SnappingClient);
 };
